+++ clang/docs/ReleaseNotes.rst
@@ -279,6 +279,8 @@ AST Matchers
 
 clang-format
 ------------
+- Add ``NumericLiteralCase`` option for for enforcing character case in
+  numeric literals.
 
 libclang
 --------